Split decision logic from analytics
